shared/execution/requirements.txt
anthropic
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
fastapi
uvicorn
requests
